Move mapping functions to seperate python module
